refactor: extract ReactableApi interface
