+++ gravitee-apim-gateway/gravitee-apim-gateway-services/gravitee-apim-gateway-services-endpoint-discovery/src/main/java/io/gravitee/gateway/services/endpoint/discovery/verticle/EndpointDiscoveryVerticle.java
@@ -117,7 +117,7 @@ public class EndpointDiscoveryVerticle extends AbstractVerticle implements Event
     }
 
     private void lookupForServiceDiscovery(Api api) {
-        if (api.isEnabled()) {
+        if (api.enabled()) {
             for (EndpointGroup group : api.getDefinition().getProxy().getGroups()) {
                 EndpointDiscoveryService discoveryService = group.getServices().get(EndpointDiscoveryService.class);
                 if (discoveryService != null && discoveryService.isEnabled()) {
